Remove maximized_{horizontal,vertical} from C code
Having these partial maximisation states is not really useful. The only
reason why we implement this seems to be EWMH. And this seems to cause
some problems with awful.placement. So, just drop this.

Setters/getters which print deprecation warnings and then just access
the "maximized" property are left behind.

Signed-off-by: Uli Schlachter <psychon@znc.in>