Travis: merge BUILD_APIDOC and DO_CHECKQA
ldoc was not installed for the DO_CHECKQA build.

This merges it with the BUILD_APIDOC build job, which is one of the
faster ones; installing ldoc into the DO_COVERAGE=coveralls would slow
that one down a lot.

This also moves the two slowest builds (with coverage reporting) to the
top, so they get started first.

Fixes https://github.com/awesomeWM/awesome/issues/1710.
Closes https://github.com/awesomeWM/awesome/pull/1716.