Add code for more direct integration with LuaRocks
This module adds LuaRocks' search path to Lua's search path if require()
fails and then tries again. Thus, this should make awesome automatically
use an lgi version that was installed through LuaRocks, even in the user
rock tree. And that's also what we are using it for: To load lgi.

Signed-off-by: Uli Schlachter <psychon@znc.in>