Fix an old todo in the docs
This new paragraph might not be great and it certainly is less
informative than what the wiki said here (it suggested to use some
library for widgets and explained how to use it). However, this new
paragraph is definitely better than the todo that was here for far too
long.

Signed-off-by: Uli Schlachter <psychon@znc.in>