awesomeConfig.cmake: GENERATE_DOC/GENERATE_MANPAGES: default to OFF
Generating documentation is slow, since it runs/configures all the
example tests, and the typical user does not care about generating the
docs when building.

Generating man pages is also rather slow, and is typically not needed.
For CMAKE_BUILD_TYPE=RELEASE generating man pages is still enabled by default.

This adds fallback `ldoc` and `man` targets to explain how to enable it,
with wrapping targets in our main Makefile.