Add a module for reading icon_theme.cache files
Since the file format is undocumented, I read the source code of Gtk's
gtk/updateiconcache.c to figure out how this works. A quick description
of the file format can be found in the source code. A quick description
for what these files actually do should be available as the API docs of
this new module. :-)

Signed-off-by: Uli Schlachter <psychon@znc.in>