Add new gears.sort module
After talking to @Elv13 about some things waiting for v4.3+, he mentioned
for #1487 a topological sort is needed.

This adds a new gears module, sort, to the library. This is pulled from
https://github.com/bungle/lua-resty-tsort, which is under the BSD license,
and slightly modified to conform to AwesomeWM code style while adding some
documentation comments.

For future sorting algorithms, if we wish to add them, I've made sort
return a table instead of just having a `gears.tsort`.