keygrabber: Turn into an object.
Previously, the capi.keygrabber was global with some kind of incomplete
wrapper called awful.keygrabber that would attempt to turn it into a
self contained object.

The API was also primitive and hard to use. Given the use case for the
keygrabber is almost limited to transactions or sequences, then the API
should focus on making this trivial and idiot proof.